Rework how aiohttp done.
I've shifted all the aiohttp into one method, dispatch_salt_message.  I've
stopped trying to use any results from that session outside of the context.
I was getting lots of errors with the ClientSession being closed at the
wrong time.

I wanted to try and re-use the ClientSession for the whole time.  I suspect
the salt-api server might be closing the session though.  And in any case, I
don't know how to use ClientSession if it's not within a context manager.